requirements.txt
tensorflow~=2.3.0
numpy~=1.18.5
annoy~=1.16.3
Pillow~=7.2.0
ruamel.yaml=0.16.12
